스케줄러 기능 추가

+++ src/main/java/com/example/HonBam/paymentsapi/toss/repository/PaymentInfoRepository.java
@@ -1,8 +1,6 @@
 package com.example.HonBam.paymentsapi.toss.repository;
 
 import com.example.HonBam.paymentsapi.toss.entity.PaymentInfo;
-import com.example.HonBam.paymentsapi.toss.entity.PaymentStatus;
-import org.hibernate.sql.Select;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -11,8 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-
-public interface PaymentInfoRepository extends JpaRepository<PaymentInfo, Integer> {
+public interface PaymentInfoRepository extends JpaRepository<PaymentInfo, Long> {
 
     Optional<PaymentInfo> findByOrderId(String orderId);
